Separate frontend files
