+++ front/src/components/index.ts
@@ -1,0 +1,7 @@
+import CommandsForm from './CommandsForm'
+import EnvForm from './EnvForm'
+
+export {
+  CommandsForm,
+  EnvForm
+}
